+++ DistributedDatabaseService.DAL/Student.cs
@@ -1,7 +1,14 @@
+using MongoDB.Bson;
+using MongoDB.Bson.Serialization.Attributes;
+
 namespace DistributedDatabaseService.DAL;
 
 public class Student
 {
+    [BsonId]
+    [BsonRepresentation(BsonType.ObjectId)]
+    public string? Id { get; set; }
+    
     public int StudentId { get; set; }
     public string FirstName { get; set; }
     public string LastName { get; set; }
@@ -11,8 +18,11 @@ public class Student
     
     // Зовнішній ключ до групи
     public int? GroupId { get; set; }
+    
+    [BsonIgnore]
     public Group? Group { get; set; }
     
     // Навігаційна властивість для відношення "багато до багатьох"
+    [BsonIgnore]
     public ICollection<StudentCourse> StudentCourses { get; set; }
 }