+++ DistributedDatabaseService.DAL/obj/Debug/net10.0/DistributedDatabaseService.DAL.AssemblyInfo.cs
@@ -13,7 +13,7 @@ using System.Reflection;
 [assembly: System.Reflection.AssemblyCompanyAttribute("DistributedDatabaseService.DAL")]
 [assembly: System.Reflection.AssemblyConfigurationAttribute("Debug")]
 [assembly: System.Reflection.AssemblyFileVersionAttribute("1.0.0.0")]
-[assembly: System.Reflection.AssemblyInformationalVersionAttribute("1.0.0+1e94aeb1b297f78c1d138a54fdd2e18deb616318")]
+[assembly: System.Reflection.AssemblyInformationalVersionAttribute("1.0.0+d1075bc2f1d6808b98987006a9d06880cce133f2")]
 [assembly: System.Reflection.AssemblyProductAttribute("DistributedDatabaseService.DAL")]
 [assembly: System.Reflection.AssemblyTitleAttribute("DistributedDatabaseService.DAL")]
 [assembly: System.Reflection.AssemblyVersionAttribute("1.0.0.0")]